Вроде добавил рендер функцию для комм и ввод комм.
